backend/pyproject.toml
[project]
name = "vibe4vets-backend"
version = "0.1.0"
description = "Veteran resource directory API"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlmodel>=0.0.22",
    "alembic>=1.14.0",
    "psycopg[binary]>=3.2.0",
    "pydantic-settings>=2.6.0",
    "email-validator>=2.0.0",
    "httpx>=0.28.0",
    "anthropic>=0.40.0",
    "pgvector>=0.3.0",
    "python-multipart>=0.0.18",
    "apscheduler>=3.10.0",
    "openpyxl>=3.1.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]
scraping = [
    "playwright>=1.49.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.3.0",
]
pdf = [
    "pdfplumber>=0.10.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 120
target-version = "py312"
exclude = [
    "scripts/seed_resources.py",  # Seed data with long string literals
    "scripts/seed_dmv_housing.py",  # Seed data with long string literals
    "scripts/seed_dmv_benefits.py",  # Seed data with long string literals
    "scripts/seed_dmv_food.py",  # Seed data with long string literals
    "scripts/import_discoveries.py",  # Complex data processing script
]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["B008"]  # Allow function calls in defaults (standard FastAPI pattern)

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = [
    "SIM117",  # Allow nested with statements in tests for clearer mock setup
    "F841",    # Allow unused variables in tests (often side effects)
]
"alembic/env.py" = [
    "E402",  # Imports must be after sys.path modification
]
"app/services/search.py" = [
    "E402",  # Imports must be after logging setup
]
"app/services/email.py" = [
    "E501",  # HTML template strings
]
"app/schemas/resource.py" = [
    "E501",  # Long Field descriptions
]
"app/services/analytics.py" = [
    "SIM108",  # Keep explicit if/else for clarity
]
"app/services/trust.py" = [
    "SIM108",  # Keep explicit if/else with comment
]
"scripts/fetch_hud_vash_awards.py" = [
    "SIM108",  # Keep explicit if/else for CLI handling
]

[tool.mypy]
python_version = "3.12"
# Relaxed mode for SQLModel/SQLAlchemy compatibility
strict = false
warn_return_any = false
warn_unused_ignores = true
disallow_untyped_defs = false
check_untyped_defs = true

[[tool.mypy.overrides]]
module = ["sqlmodel.*", "sqlalchemy.*"]
ignore_missing_imports = true
ignore_errors = true

[[tool.mypy.overrides]]
# Services and admin use SQLModel/SQLAlchemy dynamic attributes (desc, contains, search_vector)
module = [
    "app.services.search",
    "app.services.trust",
    "app.services.review",
    "app.services.resource",
    "app.services.health",
    "app.services.analytics",
    "app.api.v1.admin",
    "app.api.v1.analytics",
    "app.api.v1.feedback",
    "app.api.v1.chat",
]
ignore_errors = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
